requirements.txt
requests==2.28.1
boto3==1.26.137
psycopg2-binary==2.9.6
python-dotenv==0.21.1
apache-airflow==2.5.0
celery==5.2.7
redis==4.5.4
streamlit
plotly
sqlalchemy
pandas